Add code to try and reproduce the issue
This code is supposed to reproduce [this
issue](https://github.com/rails/importmap-rails/issues/137), so I was
expecting the stimulus controller to not work.

But, when I load the page, `hello_controller.js` is doing exactly what
it should - replacing the page content and logging to the console.

So, I haven't yet got a minimal demonstration of this problem.